Initial import for class to yaml conversion
As there are many streams and function and even more data items, adding them to the package is quite code intensive. There is a lot of duplicate code for these classes.

A first step was made when auto generating the classes from the YAML files. The classes don't need to be programmed/changed in code but just the definition in the YAML files needs to be changed. After that the classes are generated using a script.

But the better solution would be to actually read the definitions from the yaml file at runtime, instead of generating all this duplicate code. This is quite a big rework, which a lot of is already done in this commit.