Fix SECS I issues
When a block with a data length of 0 was received, the message was created without any blocks. This lead to problems using this message.

The flag for last block was set incorrectly for partial messages. The flag was set for the first block, even though it was not the last block. This lead to messages being dispatched before they were complete.