Change initialization order of GemEquipmentHandler capabilities
As the state model capability requires variables of the collection event capability, the initialization order was changed.
The collection event capability class was moved below the state model capability class in the parent class list. This way the constructor of the CollectionEventCapability is run before the StateModelsCapability, setting up required attributes for the latter.