Add SFDL
The data item definition for functions is quite complex and doesn't easily support all nesting cases. It is based on python types.
To simplify defining functions, a custom definition language is introduced, SECS Function Definition Language.
This language is based on SML to have a common base between other SECS description languages. This should simplify the definition of functions.
For details please check the documentation, a new page was added there.