Add support for SFDL in yaml function generator
To support the SFDL notation in the yaml files for the function generator, the handling of the structure data requires some changes. Same is true for the extraction of the used data items from the structure.
The template doesn't need to generate the data item imports, when the SFDL notation is used.