Add dataclasses for handling data item and function descriptors from yaml
Add a dataclass for every level of the data items and functions yaml files structure. These classes have convenience functions for loading the structure from yaml.

This includes moving the yaml files and the schemas for both data items and functions to separate files in the /secsgem/secs directory.

Add variables with the path of the new files In the modules for data items and function descriptors. (default_yaml_path and schema_path)

Update the generator code to use the schema and yaml paths from the module.